add auth.js
