+++ controllers/users.js
@@ -1,29 +1,8 @@
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
-
 const { User } = require("../models/user");
 const { HttpError, sendEmail } = require("../helpers");
 const { ctrlWrapper } = require("../decorators");
 
-const register = async (req, res) => {
-  const { email, password } = req.body;
-  const user = await User.findOne({ email });
-  const hashPassword = await bcrypt.hash(password, 10);
 
-  if (user) {
-    throw HttpError(409, "Email already in use");
-  }
-
-  const newUser = await User.create({
-    ...req.body,
-    password: hashPassword,
-  });
-
-  res.status(201).json({
-    name: newUser.name,
-    email: newUser.email,
-  })
-};
 
 const theme = async (req, res) => {
   const { id } = req.user;
@@ -51,7 +30,6 @@ const help = async (req, res) => {
 };
 
 module.exports = {
-  register: ctrlWrapper(register),
   theme: ctrlWrapper(theme),
   help: ctrlWrapper(help),
 };
